+++ stobu/core/instructioner.py
@@ -7,6 +7,7 @@
 from stobu.syss import messages as msg
 from stobu.tools.translater import translate_tags_str
 from stobu.types.action import ActDataType, ActionRecord, ActionsData, ActType
+from stobu.types.action import NORMAL_ACTIONS
 from stobu.utils.log import logger
 
 
@@ -19,19 +20,6 @@ __all__ = (
 PROC = 'INSTRUCTION'
 
 
-NORMAL_ACTIONS = [
-        ActType.BE,
-        ActType.COME,
-        ActType.DO,
-        ActType.DRAW,
-        ActType.EXPLAIN,
-        ActType.GO,
-        ActType.OCCUR,
-        ActType.TALK,
-        ActType.THINK,
-        ActType.VOICE,
-        ]
-
 INST_PARAGRAPH_START = ('P', 'pr')
 
 INST_BREAK = ('B', 'BR', 'br')
@@ -40,10 +28,15 @@ INST_PARAGRAPH_END = ('PE', 'pend')
 
 INST_ALIAS = ('A', 'alias')
 
+INST_FORESHADOW = ('FS', 'FLAG', 'foreshadow')
+
+INST_PAYOFF = ('PO', 'DISFLAG', 'payoff')
+
 
 # Main
-def actions_data_apply_instructions(actions_data: ActionsData) -> ActionsData:
+def actions_data_apply_instructions(actions_data: ActionsData, tags: dict) -> ActionsData:
     assert isinstance(actions_data, ActionsData)
+    assert isinstance(tags, dict)
 
     logger.debug(msg.PROC_START.format(proc=PROC))
 
@@ -63,6 +56,10 @@ def actions_data_apply_instructions(actions_data: ActionsData) -> ActionsData:
             elif record.subject in INST_ALIAS:
                 short, origin = record.outline.split('=')
                 alias[short] = origin
+            elif record.subject in INST_FORESHADOW:
+                tmp.append(_record_as_foreshadow_from(record, tags))
+            elif record.subject in INST_PAYOFF:
+                tmp.append(_record_as_payoff_from(record, tags))
             else:
                 logger.warning(msg.ERR_FAIL_INVALID_DATA.format(data=f"instruction type in {PROC}"))
                 continue
@@ -109,3 +106,39 @@ def _get_record_as_paragraph_end() -> ActionRecord:
 
 def _get_record_as_paragraph_start() -> ActionRecord:
     return ActionRecord(ActType.DATA, ActDataType.PARAGRAPH_START, '')
+
+
+def _record_as_foreshadow_from(record: ActionRecord, tags: dict) -> ActionRecord:
+    assert isinstance(record, ActionRecord)
+    assert isinstance(tags, dict)
+
+    subject, flag = '', record.outline
+
+    if ':' in record.outline:
+        subject, flag = record.outline.split(':')
+        if subject:
+            subject = translate_tags_str(subject, tags, True, None)
+
+    return ActionRecord(
+            ActType.DATA,
+            ActDataType.FORESHADOW,
+            subject,
+            flag)
+
+
+def _record_as_payoff_from(record: ActionRecord, tags: dict) -> ActionRecord:
+    assert isinstance(record, ActionRecord)
+    assert isinstance(tags, dict)
+
+    subject, flag = '', record.outline
+
+    if ':' in record.outline:
+        subject, flag = record.outline.split(':')
+        if subject:
+            subject = translate_tags_str(subject, tags, True, None)
+
+    return ActionRecord(
+            ActType.DATA,
+            ActDataType.PAYOFF,
+            subject,
+            flag)